Make sure we are following : https://zuelligpharma.atlassian.net/wiki/spaces/DDPRG/pages/231670169/Pull+Request+PR+reviews

### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-24790

### TDD PR url
- None
Note: I think it just passes the value from the contractor and there is already a unit test to check so there is no need for additional test cases
![Screenshot 2024-07-24 at 10.59.59.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/113223/attachments/Screenshot%202024-07-24%20at%2010.59.59.png)

### Root cause
- Missing unit price field when copy with from contractor.

### Last related PR
- None

### Solution
- Add unit price field when copy with from contractor.

### Test account
- environment: uat
- username: namthanhnguyen

### Output screenshot
- ![Screenshot 2024-07-24 at 10.30.29.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/113223/attachments/Screenshot%202024-07-24%20at%2010.30.29.png)

EZRX-24790: update unit price in return submitted

+++ lib/domain/returns/entities/return_material.dart
@@ -88,7 +88,7 @@ class ReturnMaterial with _$ReturnMaterial {
         itemNumber: itemNumber,
         assignmentNumber: assignmentNumber,
         batch: batch,
-        unitPrice: unitPrice,
+        unitPrice: unitPrice.doubleValue,
         balanceQty: balanceQuantity,
       );
